requirements.txt
Flask>=2.1.1
cryptography==36.0.1
PyJWT==2.3.0
pytest==7.0.1
requests==2.27.1
